refactor(semantic): share call graph construction
